requirements.txt
django_haystack==3.3.0
Flask==1.1.2
psutil==5.9.8
